Refactor: Consolidate redundant TypeScript type definitions
Summary of changes:
- Centralized `Gender` in `types/core.ts` as uppercase string literals ('MALE' | 'FEMALE' | 'NEUTRAL').
- Consolidated `WorkoutStatus` in `types/workout.ts` to include 'idle', 'running', 'paused', 'finished'.
- Unified Spotify integration types by promoting `SpotifyPlaylist` and `SpotifyPlaylistItem` in `types/core.ts` as canonical definitions.
- Centralized `SpotifyCommand` and `SpotifyTokenResponse` in `types/core.ts`.
- Standardized heart rate metric naming conventions to use `value` (for BPM) and `percentage` (for % of Max HR) across UI components (`HrTile`, `HrTileWrapper`), hooks, and data structures.
- Deleted obsolete type definition files: `types/user.ts`, `types/service.ts`, `types/heart-rate.ts`, and `types/spotify.ts`.
- Updated all affected components (`PlaylistSelector`, `PlaylistDetails`, `HrTile`, `ConnectView`), services (`SpotifyTokenManager`, `SpotifyPolling`), and tests to align with the new canonical types.
- Resolved build issues and verified changes with 100% passing unit tests.

These changes improve consistency, reduce maintenance overhead, and establish `types/core.ts` as the single source of truth for the application's domain models.
