+++ app/api/spotify/playlists/route.ts
@@ -5,7 +5,6 @@
 import { NextResponse, NextRequest } from 'next/server'
 import { withErrorHandler } from '@/lib/middleware/errorHandler'
 import { getAuthenticatedSpotifyApi } from '@/lib/spotify/sdk'
-import { UserPlaylistDto } from '@/types/spotify'
 import { SpotifyPlaylist } from '@/types/core'
 
 /**
@@ -33,7 +32,7 @@ async function getPlaylists(_req: NextRequest) {
   ]
 
   // 6. Map user playlists to include full data (images, descriptions, track counts, etc.)
-  const userPlaylists: UserPlaylistDto[] = playlistsResponse.items.map(
+  const userPlaylists: SpotifyPlaylist[] = playlistsResponse.items.map(
     (playlist) => ({
       id: playlist.id,
       name: playlist.name,
